Split requests and responses
