+++ src/main/java/br/com/fiap/techchallenge/lanchonete/adapters/web/mappers/ProdutoMapper.java
@@ -1,6 +1,6 @@
 package br.com.fiap.techchallenge.lanchonete.adapters.web.mappers;
 
-import br.com.fiap.techchallenge.lanchonete.adapters.web.models.ProdutoResponse;
+import br.com.fiap.techchallenge.lanchonete.adapters.web.models.responses.ProdutoResponse;
 import br.com.fiap.techchallenge.lanchonete.core.dtos.ProdutoDTO;
 import org.springframework.stereotype.Component;
