+++ src/main/java/br/com/fiap/techchallenge/lanchonete/adapters/web/models/responses/ProdutoResponse.java
@@ -1,4 +1,4 @@
-package br.com.fiap.techchallenge.lanchonete.adapters.web.models;
+package br.com.fiap.techchallenge.lanchonete.adapters.web.models.responses;
 
 import br.com.fiap.techchallenge.lanchonete.core.domain.entities.enums.CategoriaEnum;
